Rename /ws to /ws-roster
gh-395

And connect to connect-roster.
